requirements.txt
bs4
dash
numpy
Flask
plotly
pandas
requests
gunicorn
pandas-ta
python-dotenv
python-binance
python-telegram-bot
